requirements.txt
streamlit
python-dotenv
openai>=1.0.0
faiss-cpu
sentence-transformers
PyMuPDF
numpy
